refactor: move MediaPlayer, InlineError, ErrorIcon to components
