+++ packages/ai-chat/src/chat/components-legacy/MessageComponent.tsx
@@ -22,7 +22,7 @@ import React, { KeyboardEvent, PureComponent } from "react";
 
 import { nodeToText } from "../utils/domUtils";
 import { Avatar } from "./Avatar";
-import { InlineError } from "./responseTypes/error/InlineError";
+import { InlineError } from "../components/util/InlineError";
 import VisuallyHidden from "../components/util/VisuallyHidden";
 
 // Inline helper components (previously in util/IconHolder.tsx and util/ImageWithFallback.tsx)
