+++ packages/ai-chat/src/chat/components-legacy/responseTypes/_imports.scss
@@ -12,7 +12,6 @@
 @use "carousel/imports" as responseType-carousel;
 @use "conversationalSearch/imports" as responseType-conversationalSearch;
 @use "datePicker/imports" as responseType-datePicker;
-@use "error/imports" as responseType-error;
 @use "grid/imports" as responseType-grid;
 @use "iframe/imports" as responseType-iframe;
 @use "image/imports" as responseType-image;
