+++ packages/ai-chat/src/chat/components-legacy/responseTypes/conversationalSearch/ConversationalSearch.tsx
@@ -1,5 +1,5 @@
 /*
- *  Copyright IBM Corp. 2025
+ *  Copyright IBM Corp. 2025, 2026
  *
  *  This source code is licensed under the Apache-2.0 license found in the
  *  LICENSE file in the root directory of this source tree.
@@ -13,7 +13,7 @@ import { ScrollElementIntoViewFunction } from "../../MessagesComponent";
 import { useLanguagePack } from "../../../hooks/useLanguagePack";
 import { LocalMessageItem } from "../../../../types/messaging/LocalMessageItem";
 import { SkeletonPlaceholder } from "../../../components/util/SkeletonPicker";
-import InlineError from "../error/InlineError";
+import InlineError from "../../../components/util/InlineError";
 import { CitationCard } from "../util/citations/CitationCard";
 import { ConversationalSearchText } from "./ConversationalSearchText";
 import {
