+++ packages/ai-chat/src/chat/components-legacy/responseTypes/custom/UserDefinedResponse.tsx
@@ -1,5 +1,5 @@
 /*
- *  Copyright IBM Corp. 2025
+ *  Copyright IBM Corp. 2025, 2026
  *
  *  This source code is licensed under the Apache-2.0 license found in the
  *  LICENSE file in the root directory of this source tree.
@@ -18,7 +18,7 @@ import React from "react";
 
 import { HasServiceManager } from "../../../hocs/withServiceManager";
 import { useLanguagePack } from "../../../hooks/useLanguagePack";
-import InlineError from "../error/InlineError";
+import InlineError from "../../../components/util/InlineError";
 
 interface UserDefinedResponseProps extends HasServiceManager {
   /**
