--- packages/ai-chat/src/chat/components-legacy/responseTypes/error/_imports.scss
@@ -1,8 +1,0 @@
-/* 
- *  Copyright IBM Corp. 2025
- *  
- *  This source code is licensed under the Apache-2.0 license found in the
- *  LICENSE file in the root directory of this source tree.
- */
-
-@use "InlineError";
