+++ packages/ai-chat/src/chat/components-legacy/responseTypes/iframe/IFrameComponent.tsx
@@ -1,5 +1,5 @@
 /*
- *  Copyright IBM Corp. 2025
+ *  Copyright IBM Corp. 2025, 2026
  *
  *  This source code is licensed under the Apache-2.0 license found in the
  *  LICENSE file in the root directory of this source tree.
@@ -14,7 +14,7 @@ import { useAriaAnnouncer } from "../../../hooks/useAriaAnnouncer";
 import { useLanguagePack } from "../../../hooks/useLanguagePack";
 import { RESPONSE_TYPE_TIMEOUT_MS } from "../../../utils/constants";
 import { MountChildrenOnDelay } from "../../../components/util/MountChildrenOnDelay";
-import InlineError from "../error/InlineError";
+import InlineError from "../../../components/util/InlineError";
 
 interface IFrameComponentProps {
   /**
