+++ packages/ai-chat/src/chat/components-legacy/responseTypes/iframe/InlineIFrame.tsx
@@ -1,5 +1,5 @@
 /*
- *  Copyright IBM Corp. 2025
+ *  Copyright IBM Corp. 2025, 2026
  *
  *  This source code is licensed under the Apache-2.0 license found in the
  *  LICENSE file in the root directory of this source tree.
@@ -17,7 +17,7 @@ import {
 } from "../../../utils/cspStyleUtils";
 import { getMediaDimensions } from "../../../utils/messageUtils";
 import { getResponsiveElementPaddingValue } from "../../../utils/miscUtils";
-import InlineError from "../error/InlineError";
+import InlineError from "../../../components/util/InlineError";
 import { IFrameComponent } from "./IFrameComponent";
 import { IFrameItem } from "../../../../types/messaging/Messages";
