+++ packages/ai-chat/src/chat/components-legacy/responseTypes/util/StreamingRichText.tsx
@@ -1,5 +1,5 @@
 /*
- *  Copyright IBM Corp. 2025
+ *  Copyright IBM Corp. 2025, 2026
  *
  *  This source code is licensed under the Apache-2.0 license found in the
  *  LICENSE file in the root directory of this source tree.
@@ -11,7 +11,7 @@ import React from "react";
 
 import { useLanguagePack } from "../../../hooks/useLanguagePack";
 import { LocalMessageItemStreamingState } from "../../../../types/messaging/LocalMessageItem";
-import InlineError from "../error/InlineError";
+import InlineError from "../../../components/util/InlineError";
 import { MarkdownWithDefaults } from "../../../components/util/MarkdownWithDefaults";
 import { TextItem } from "../../../../types/messaging/Messages";
