+++ packages/ai-chat/src/chat/components-legacy/responseTypes/util/_imports.scss
@@ -1,5 +1,5 @@
 /* 
- *  Copyright IBM Corp. 2025
+ *  Copyright IBM Corp. 2025, 2026
  *  
  *  This source code is licensed under the Apache-2.0 license found in the
  *  LICENSE file in the root directory of this source tree.
@@ -9,7 +9,6 @@
 @use "ClickableImage";
 @use "Description";
 @use "FooterButtonComponents";
-@use "MediaPlayer";
 @use "Metablock";
 @use "TextHolderTile";
 @use "SearchResultBody";
