+++ packages/ai-chat/src/chat/components-legacy/responseTypes/video/VideoComponent.tsx
@@ -1,5 +1,5 @@
 /*
- *  Copyright IBM Corp. 2025
+ *  Copyright IBM Corp. 2025, 2026
  *
  *  This source code is licensed under the Apache-2.0 license found in the
  *  LICENSE file in the root directory of this source tree.
@@ -10,7 +10,10 @@
 import React from "react";
 
 import { HasBaseHeight } from "../../../../types/utilities/HasBaseHeight";
-import { MediaPlayer, MediaPlayerContentConfig } from "../util/MediaPlayer";
+import {
+  MediaPlayer,
+  MediaPlayerContentConfig,
+} from "../../../components/util/MediaPlayer";
 import { MessageResponseTypes } from "../../../../types/messaging/Messages";
 
 type VideoComponentConfig = HasBaseHeight;
