+++ packages/ai-chat/src/chat/components/util/ErrorIcon.tsx
@@ -1,5 +1,5 @@
 /*
- *  Copyright IBM Corp. 2025
+ *  Copyright IBM Corp. 2025, 2026
  *
  *  This source code is licensed under the Apache-2.0 license found in the
  *  LICENSE file in the root directory of this source tree.
@@ -8,11 +8,11 @@
  */
 
 import ErrorFilled16 from "@carbon/icons/es/error--filled/16.js";
-import { carbonIconToReact } from "../utils/carbonIcon";
+import { carbonIconToReact } from "../../utils/carbonIcon";
 import cx from "classnames";
 import React from "react";
 
-import { HasClassName } from "../../types/utilities/HasClassName";
+import { HasClassName } from "../../../types/utilities/HasClassName";
 
 const ErrorFilled = carbonIconToReact(ErrorFilled16);
