+++ packages/ai-chat/src/chat/components/util/InlineError.tsx
@@ -1,5 +1,5 @@
 /*
- *  Copyright IBM Corp. 2025
+ *  Copyright IBM Corp. 2025, 2026
  *
  *  This source code is licensed under the Apache-2.0 license found in the
  *  LICENSE file in the root directory of this source tree.
@@ -9,9 +9,9 @@
 
 import React from "react";
 
-import { useLanguagePack } from "../../../hooks/useLanguagePack";
-import { ErrorIcon } from "../../ErrorIcon";
-import { MarkdownWithDefaults } from "../../../components/util/MarkdownWithDefaults";
+import { useLanguagePack } from "../../hooks/useLanguagePack";
+import { ErrorIcon } from "./ErrorIcon";
+import { MarkdownWithDefaults } from "./MarkdownWithDefaults";
 
 export function InlineError({ text }: { text?: string }) {
   const languagePack = useLanguagePack();
